add all classes

+++ module_3/src/main/java/ua/com/alevel/service/impl/CategoryServiceImpl.java
@@ -2,20 +2,24 @@ package ua.com.alevel.service.impl;
 
 import org.springframework.stereotype.Service;
 import ua.com.alevel.persistence.dao.CategoryDao;
+import ua.com.alevel.persistence.dao.TransactionDao;
 import ua.com.alevel.persistence.datatable.DataTableRequest;
 import ua.com.alevel.persistence.datatable.DataTableResponse;
-import ua.com.alevel.persistence.entity.Account;
 import ua.com.alevel.persistence.entity.Category;
 import ua.com.alevel.service.CategoryService;
 import ua.com.alevel.util.WebResponseUtil;
+
+import java.util.List;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryDao categoryDao;
+    private final TransactionDao transactionDao;
 
-    public CategoryServiceImpl(CategoryDao categoryDao) {
+    public CategoryServiceImpl(CategoryDao categoryDao, TransactionDao transactionDao) {
         this.categoryDao = categoryDao;
+        this.transactionDao = transactionDao;
     }
 
     @Override
@@ -30,7 +34,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
-        categoryDao.delete(id);
+        Category category = categoryDao.findById(id);
+        if(!transactionDao.isExistCategoryInTransactions(category)) {
+            try {
+                categoryDao.delete(id);
+            } catch (Exception e) {
+                System.out.println("Cannot delete category");
+            }
+        }
     }
 
     @Override
@@ -44,5 +55,10 @@ public class CategoryServiceImpl implements CategoryService {
         long count = categoryDao.count();
         WebResponseUtil.initDataTableResponse(request, dataTableResponse, count);
         return dataTableResponse;
+    }
+
+    @Override
+    public List<Category> findMapCategories() {
+        return categoryDao.findMapCategories();
     }
 }
